Refactor scout models
