+++ cg/apps/scout/scoutapi.py
@@ -7,9 +7,8 @@ from subprocess import CalledProcessError
 from typing import List, Optional
 
 import yaml
-
 from cg.apps.scout.scout_export import ScoutExportCase, Variant
-from cg.meta.upload.scout.scout_load_config import ScoutLoadConfig
+from cg.models.scout.scout_load_config import ScoutLoadConfig
 from cg.utils.commands import Process
 
 LOG = logging.getLogger(__name__)
