+++ cg/models/scout/scout_load_config.py
