+++ cg/store/models.py
@@ -1,10 +1,7 @@
-# -*- coding: utf-8 -*-
 import datetime as dt
 from typing import List
 
 import alchy
-from sqlalchemy import Column, ForeignKey, Table, UniqueConstraint, orm, types
-
 from cg.constants import (
     CASE_ACTIONS,
     FLOWCELL_STATUS,
@@ -16,6 +13,7 @@ from cg.constants import (
     DataDelivery,
     Pipeline,
 )
+from sqlalchemy import Column, ForeignKey, Table, UniqueConstraint, orm, types
 
 Model = alchy.make_declarative_base(Base=alchy.ModelBase)
